refactor: Regroup component related to Identity together
