+++ packages/cozy-sharing/src/components/Identity/Identity.jsx
