+++ packages/cozy-sharing/src/components/Identity/Identity.stories.jsx
@@ -1,6 +1,7 @@
 import Identity from './Identity'
 
 const meta = {
+  title: 'Components/Identity/Identity',
   component: Identity,
   args: {}
 }
