+++ packages/cozy-sharing/src/components/Identity/OwnerIdentity.jsx
