+++ packages/cozy-sharing/src/components/Identity/RecipientPlusX.stories.jsx
