+++ packages/cozy-sharing/src/components/Identity/RecipientWithoutStatus.jsx
