+++ packages/cozy-sharing/src/components/Identity/RecipientWithoutStatus.stories.jsx
